api/requirements.txt
Flask==1.1.2
Jinja2==2.11.3
MarkupSafe==1.1.1
SQLAlchemy==1.4.5
Werkzeug==1.0.1
click==7.1.2
flask-sqlalchemy==2.5.1
greenlet==1.0.0
itsdangerous==1.1.0
yoyo_migrations==7.3.1
pytest